+++ emg3d/solver.py
@@ -32,13 +32,13 @@ from dataclasses import dataclass
 from emg3d import utils
 from emg3d import njitted
 
-__all__ = ['solver', 'multigrid', 'smoothing', 'restriction', 'prolongation',
+__all__ = ['solve', 'multigrid', 'smoothing', 'restriction', 'prolongation',
            'residual', 'krylov', 'MGParameters', 'RegularGridProlongator']
 
 
 # MAIN USER-FACING FUNCTION
-def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
-           semicoarsening=False, linerelaxation=False, verb=2, **kwargs):
+def solve(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
+          semicoarsening=False, linerelaxation=False, verb=2, **kwargs):
     r"""Solver for 3D CSEM data with tri-axial electrical anisotropy.
 
     The principal solver of `emg3d` is using the multigrid method as presented
@@ -120,10 +120,6 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
               :func:`scipy.sparse.linalg.bicgstab`;
             - 'cgs': Conjugate Gradient Squared
               :func:`scipy.sparse.linalg.cgs`;
-            - 'gmres': Generalized Minimal RESidual
-              :func:`scipy.sparse.linalg.gmres`;
-            - 'lgmres': Improvement of GMRES using alternating residual
-              vectors :func:`scipy.sparse.linalg.lgmres`;
             - 'gcrotmk': GCROT: Generalized Conjugate Residual with inner
               Orthogonalization and Outer Truncation
               :func:`scipy.sparse.linalg.gcrotmk`.
@@ -175,6 +171,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         - 2: Print runtime and information about the method.
         - 3: Print additional information for each MG-cycle.
         - 4: Print everything (slower due to additional error calculations).
+        - -1: Print one-liner (dynamically updated).
 
     **kwargs : Optional solver options:
 
@@ -247,6 +244,8 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         - ``it_mg``: Number of multigrid iterations;
         - ``it_ssl``: Number of SSL iterations;
         - ``time``: Runtime (s).
+        - ``runtime_at_cycle``: Runtime after each cycle (s).
+        - ``error_at_cycle``: Absolute error after each cycle.
 
 
     Examples
@@ -265,7 +264,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
     >>> sfield = emg3d.utils.get_source_field(
     >>>         grid, src=[4, 4, 4, 0, 0], freq=10)
     >>> # Calculate the electric signal.
-    >>> efield = emg3d.solver.solver(grid, model, sfield, verb=3)
+    >>> efield = emg3d.solve(grid, model, sfield, verb=3)
     >>> # Get the corresponding magnetic signal.
     >>> hfield = emg3d.utils.get_h_field(grid, model, efield)
     .
@@ -313,6 +312,7 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
 
     # Calculate reference error for tolerance.
     var.l2_refe = njitted.l2norm(sfield)
+    var.error_at_cycle[0] = var.l2_refe
 
     # Check sfield and get efield
     if sfield.freq is None:
@@ -380,17 +380,19 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         multigrid(grid, vmodel, sfield, efield, var)
 
     # Print runtime information.
-    if var.sslsolver:  # sslsolver-specific info.
-        info = f"   > Solver steps     : {var._ssl_it}\n"
-        if var.cycle:
-            info += f"   > MG prec. steps   : {var.it}\n"
-    elif var.cycle:    # multigrid-specific info.
-        info = f"   > MG cycles        : {var.it}\n"
-    info += f"   > Final rel. error : {var.l2/var.l2_refe:.3e}\n\n"  # Error.
-    info += f":: emg3d END   :: {var.time.now} :: "  # END and time.
-    time = var.time.runtime
-    info += f"runtime = {time}\n"                    # Total runtime.
-    var.cprint(info, 1)
+    if var.verb < 0:
+        var.one_liner(var.l2, True)
+    elif var.verb > 1:
+        if var.sslsolver:  # sslsolver-specific info.
+            info = f"   > Solver steps     : {var._ssl_it}\n"
+            if var.cycle:
+                info += f"   > MG prec. steps   : {var.it}\n"
+        elif var.cycle:    # multigrid-specific info.
+            info = f"   > MG cycles        : {var.it}\n"
+        info += f"   > Final rel. error : {var.l2/var.l2_refe:.3e}\n\n"
+        info += f":: emg3d END   :: {var.time.now} :: "
+        info += f"runtime = {var.time.runtime}\n"
+        var.cprint(info, 1)
 
     # Assemble the info_dict if return_info
     if var.return_info:
@@ -404,7 +406,9 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
             'tol': var.tol,              # Tolerance (abs_error<ref_error*tol).
             'it_mg': var.it,             # Multigrid iterations.
             'it_ssl': var._ssl_it,       # SSL iterations.
-            'time': time.seconds,        # Runtime (s).
+            'time': var.runtime_at_cycle[-1],          # Runtime (s).
+            'runtime_at_cycle': var.runtime_at_cycle,  # Runtime at cycle (s).
+            'error_at_cycle': var.error_at_cycle,      # Abs. error at cycle.
         }
 
     # Return depending on input arguments; or nothing.
@@ -414,6 +418,19 @@ def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
         return efield
     elif var.return_info:                  # info.
         return info_dict
+
+
+def solver(grid, model, sfield, efield=None, cycle='F', sslsolver=False,
+           semicoarsening=False, linerelaxation=False, verb=2, **kwargs):
+    """Alias to solve(), for backwards compatibility."""
+
+    # Issue warning for backwards compatibility.
+    print("\n* WARNING :: ``emg3d.solver.solver()`` is renamed to "
+          "``emg3d.solve()``.\n             Use the new ``emg3d.solve()``, as "
+          "``solver()`` will be\n             removed in the future.")
+
+    return solve(grid, model, sfield, efield, cycle, sslsolver, semicoarsening,
+                 linerelaxation, verb, **kwargs)
 
 
 # SOLVERS
@@ -473,7 +490,7 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
     l2_last = residual(grid, model, sfield, efield, True)
 
     # Keep track on the levels during the first cycle, for QC.
-    if var._first_cycle:
+    if var._first_cycle and var.verb > 2:
         var._level_all.append(level)
 
     # Print initial call info.
@@ -540,7 +557,7 @@ def multigrid(grid, model, sfield, efield, var, **kwargs):
             prolongation(grid, efield, cgrid, cefield, sc_dir)
 
             # Append current prolongation level for QC.
-            if var._first_cycle:
+            if var._first_cycle and var.verb > 2:
                 var._level_all.append(level)
 
             # (B.5) Post-smoothing (nu_post).
@@ -692,39 +709,44 @@ def krylov(grid, model, sfield, efield, var):
         # Update iteration count.
         var._ssl_it += 1
 
-        # Calculate and print error (only if verbose).
-        if var.verb > 2:
+        # Add current runtime and error to var.
+        var.runtime_at_cycle = np.r_[var.runtime_at_cycle, var.time.elapsed]
+        var.l2 = residual(grid, model, sfield, utils.Field(grid, x), True)
+        var.error_at_cycle = np.r_[var.error_at_cycle, var.l2]
 
-            # 'gmres' returns the error, not the solution, in the callback.
-            if var.sslsolver == 'gmres':
-                var.l2 = x
-            else:
-                var.l2 = residual(
-                        grid, model, sfield, utils.Field(grid, x), True)
+        # Print error (only if verbose).
+        if var.verb > 2:
 
             log = f"   [{var.time.now}]   {var.l2/var.l2_refe:.3e} "
             log += f" after {var._ssl_it:3} {var.sslsolver}-cycles"
 
             # For those solvers who run an iteration before the first
-            # preconditioner run ['lgmres', 'gcrotmk'].
+            # preconditioner run ['gcrotmk'].
             if var._ssl_it == 1 and var.it == 0 and var.cycle is not None:
                 log += "\n"
 
             var.cprint(log, 2)
 
-    # Solve the system with sslsolver.
-    efield.field, i = getattr(ssl, var.sslsolver)(
-            A=A, b=sfield, x0=efield, tol=var.tol, maxiter=var.ssl_maxit,
-            atol=1e-30, M=M, callback=callback)
+        elif var.verb < 0:
 
-    # Calculate final error, if not done in the callback.
-    if var.verb < 3:
-        var.l2 = residual(grid, model, sfield, efield, True)
+            var.one_liner(var.l2)
+
+    # Solve the system with sslsolver.
+    # The ssl solvers do not abort if the norm diverges or is not finite. We
+    # therefore throw an exception in ``_terminate``, and catch it here.
+    try:
+        efield.field, i = getattr(ssl, var.sslsolver)(
+                A=A, b=sfield, x0=efield, tol=var.tol, maxiter=var.ssl_maxit,
+                atol=1e-30, M=M, callback=callback)
+    except _ConvergenceError:
+        i = -1  # Mark it as error; returned field is all zero.
+        var.exit_message += " (returned field is zero)"
 
     # Convergence-checks for sslsolver.
     pre = "\n   > "
     if i < 0:
-        var.exit_message = f"Error in {var.sslsolver}"
+        if var.exit_message == '':
+            var.exit_message = f"Error in {var.sslsolver}"
         pre = "\n* ERROR   :: "
     elif i > 0:
         var.exit_message = "MAX. ITERATION REACHED, NOT CONVERGED"
@@ -1092,6 +1114,8 @@ class MGParameters:
         self.exit_message = ''     # For convergence status.
 
         self.time = utils.Time()   # Timer.
+        self.runtime_at_cycle = np.array([0.])  # Store runtime per cycle.
+        self.error_at_cycle = np.array([0.])    # Store error per cycle.
         self.do_return = True      # Whether or not to return the efield.
 
         # 1. Set everything related to semicoarsening and line relaxation.
@@ -1217,6 +1241,32 @@ class MGParameters:
         if self.verb > verbosity:
             print(info, **kwargs)
 
+    def one_liner(self, l2_last, last=False):
+        """Print continuously updated one-liner.
+
+        Parameters
+        ----------
+        l2_last : float
+            Current error.
+
+        last : bool
+            If True, adds ``exit_message`` and finishes line.
+
+        """
+        # Collect info.
+        info = f":: emg3d :: {l2_last/self.l2_refe:.1e}; "  # Absolute error.
+        if self.sslsolver:  # For multigrid as preconditioner.
+            info += f"{self._ssl_it}({self.it}); "
+        else:               # Stand-alone multigrid.
+            info += f"{self.it}; "
+        info += f"{self.time.runtime}"  # Runtime
+
+        # Print depending on `exit`.
+        if last:
+            print(info+f"; {self.exit_message}")
+        else:
+            print(info, end='\r')
+
     def _semicoarsening(self):
         """Set everything related to semicoarsening."""
 
@@ -1291,7 +1341,7 @@ class MGParameters:
         """Set everything related to solver and MG-cycle."""
 
         # sslsolver.
-        solvers = ['bicgstab', 'cgs', 'gmres', 'lgmres', 'gcrotmk']
+        solvers = ['bicgstab', 'cgs', 'gcrotmk']
         if self.sslsolver is True:
             self.sslsolver = 'bicgstab'
         elif self.sslsolver is not False and self.sslsolver not in solvers:
@@ -1550,11 +1600,16 @@ def _print_cycle_info(var, l2_last, l2_prev):
         Last and previous errors (l2-norms).
 
     """
+    # Add current runtime to var.
+    var.runtime_at_cycle = np.r_[var.runtime_at_cycle, var.time.elapsed]
+    var.error_at_cycle = np.r_[var.error_at_cycle, l2_last]
 
     # Start info string, return if not enough verbose.
-    if var.verb < 3:
+    if var.verb < 0:  # One-liner
+        var.one_liner(l2_last)
+        return
+    elif var.verb < 3:
         # Set first_cycle to False, to stop logging.
-        var._first_cycle = False
         return
     elif var.verb > 3:
         info = "\n"
@@ -1666,31 +1721,31 @@ def _terminate(var, l2_last, l2_prev, it):
     """
 
     finished = False
+    sslabort = False
 
-    if var.sslsolver:  # If multigrid as preconditioner, exit silently.
-        if it == var.maxit:
-            finished = True
+    if l2_last < var.tol*var.l2_refe:    # Converged.
+        var.exit_message = "CONVERGED"
+        finished = True
 
-    else:
+    elif l2_last > 10*var.l2_refe or not np.isfinite(l2_last):  # Diverged.
+        var.exit_message = "DIVERGED"
+        finished = True
+        sslabort = True  # Force abort if sslsolver.
 
-        if l2_last < var.tol*var.l2_refe:    # Converged.
-            var.exit_message = "CONVERGED"
-            finished = True
+    elif it > 2 and l2_last >= l2_prev:  # Stagnated.
+        var.exit_message = "STAGNATED"
+        finished = True
+        sslabort = True  # Force abort if sslsolver.
 
-        elif l2_last > 10*var.l2_refe or np.isnan(l2_last):  # Diverged.
-            var.exit_message = "DIVERGED"
-            finished = True
+    elif it == var.maxit:                # Maximum iterations reached.
+        var.exit_message = "MAX. ITERATION REACHED, NOT CONVERGED"
+        finished = True
 
-        elif it > 2 and l2_last >= l2_prev:  # Stagnated.
-            var.exit_message = "STAGNATED"
-            finished = True
-
-        elif it == var.maxit:                # Maximum iterations reached.
-            var.exit_message = "MAX. ITERATION REACHED, NOT CONVERGED"
-            finished = True
-
-        # Print info.
-        if finished:
+    # Force abort (ssl solver) or print info.
+    if finished:
+        if var.sslsolver and sslabort:  # Force abortion of SSL solver.
+            raise _ConvergenceError
+        elif not var.sslsolver:  # Print info (if not preconditioner).
             if var.verb < 4:
                 add = "\n"
             else:
@@ -1799,3 +1854,8 @@ def _get_prolongation_coordinates(grid, d1, d2):
     D2, D1 = np.broadcast_arrays(
             getattr(grid, 'vectorN'+d2), getattr(grid, 'vectorN'+d1)[:, None])
     return np.r_[D1.ravel('F'), D2.ravel('F')].reshape(-1, 2, order='F')
+
+
+class _ConvergenceError(Exception):
+    """Custom exception for convergence issues with SSL solvers."""
+    pass