+++ CHANGELOG.rst
@@ -2,6 +2,55 @@ Changelog
 #########
 
 
+*latest*
+--------
+
+- Rename ``solver.solver`` to ``solver.solve``; load ``solve`` also into the
+  main namespace as ``emg3d.solve``.
+- Adjustment to ``utils.get_hx_h0`` (this might change your boundaries): The
+  calculation domain is now calculated so that the distance for the signal
+  travelling from the source to the boundary and back to the most remote
+  receiver is at least two wavelengths away. If this is within the provided
+  domain, then now extra buffer is added around the domain.
+
+
+*v0.9.2* : Complex sources
+--------------------------
+
+**2019-12-26**
+
+- Strength input for ``get_source_field`` can now be complex; it also stores
+  now the source location and its strength and moment.
+- ``get_receiver`` can now take entire ``Field`` instances, and returns in that
+  case (``fx``, ``fy``, ``fz``) at receiver locations.
+- Krylov subspace solvers:
+
+  - Solver now finishes in the middle of preconditioning cycles if tolerance is
+    reached.
+  - Solver now aborts if solution diverges or stagnates also for the SSL
+    solvers; it fails and returns a zero field.
+  - Removed ``gmres`` and ``lgmres`` from the supported SSL solvers; they do
+    not work nice for this problem. Supported remain ``bicgstab`` (default),
+    ``cgs``, and ``gcrotmk``.
+
+- Various small things:
+
+  - New attribute ``Field.is_electric``, so the field knows if it is electric
+    or magnetic.
+  - New ``verb``-possibility: ``verb=-1`` is a continuously updated one-liner,
+    ideal to monitor large sets of calculations or in inversions.
+  - The returned ``info`` dictionary contains new keys:
+
+    - ``runtime_at_cycle``: accumulated total runtime at each cycle;
+    - ``error_at_cycle``: absolute error at each cycle.
+
+  - Simple ``__repr__`` for ``TensorMesh``, ``Model``, ``Fourier``, ``Time``.
+
+- Bugfixes:
+
+  - Related to ``get_hx_h0``, ``data_write``, printing in ``Fourier``.
+
+
 *v0.9.1* : VolumeModel
 ----------------------
 
@@ -13,12 +62,11 @@ Changelog
     electric permittivity.
   - ``VolumeModel`` contains the volume-averaged values eta and zeta; called
     from within ``emg3d.solver.solver``.
-
-    - Full wave equation is enabled again, via ``epsilon_r``; by default it is
-      set to None, hence diffusive approximation.
-    - Model parameters are now internally stored as 1D arrays.
-    - An {isotropic, VTI, HTI} initiated model can be changed by providing the
-      missing resistivities.
+  - Full wave equation is enabled again, via ``epsilon_r``; by default it is
+    set to None, hence diffusive approximation.
+  - Model parameters are now internally stored as 1D arrays.
+  - An {isotropic, VTI, HTI} initiated model can be changed by providing the
+    missing resistivities.
 
 - Bugfix: Up and till version 0.8.1 there was a bug. If resistivity was set
   with slices, e.g., ``model.res[:, :, :5]=1e10``, it DID NOT update the
@@ -27,10 +75,10 @@ Changelog
 
 - Various:
 
-    - The log now lists the version of emg3d.
-    - PEP8: internal imports now use absolute paths instead of relative ones.
-    - Move from conda-channel ``prisae`` to ``conda-forge``.
-    - Automatic deploy for PyPi and conda-forge.
+  - The log now lists the version of emg3d.
+  - PEP8: internal imports now use absolute paths instead of relative ones.
+  - Move from conda-channel ``prisae`` to ``conda-forge``.
+  - Automatic deploy for PyPi and conda-forge.
 
 
 *v0.9.0* : Fourier